Fix conversion of IPv4/IPv6 bindata to strings
The documentation mentions taking care of bindata (buffers in node), such as IPv4 and IPV6 addresses, and converting them to easy-to-use javascript strings where possible. This was apparently not happening for at least `just_address_answers`, due to a small bug in the `getdns_dict_to_ip_string()` function. The tests were written to match the buffer output, so this documentation mismatch/bug was not detected.

This fixes #16 and should align the documentation and the code.